- Added OrderItemParser for parsing of OrderItems  - DeliveryOrder and Order Item parsing

+++ app/src/main/java/com/goleep/driverapp/viewmodels/CashSalesViewModel.java
@@ -11,19 +11,15 @@ import com.goleep.driverapp.constants.UrlConstants;
 import com.goleep.driverapp.interfaces.NetworkAPICallback;
 import com.goleep.driverapp.interfaces.UILevelNetworkCallback;
 import com.goleep.driverapp.services.network.NetworkService;
-import com.goleep.driverapp.services.network.responsemodels.DoDetailResponseModel;
+import com.goleep.driverapp.services.network.jsonparsers.DeliveryOrderParser;
+import com.goleep.driverapp.services.network.jsonparsers.OrderItemParser;
 import com.goleep.driverapp.services.room.AppDatabase;
 import com.goleep.driverapp.services.room.RoomDBService;
 import com.goleep.driverapp.services.room.entities.DeliveryOrderEntity;
 import com.goleep.driverapp.services.room.entities.OrderItemEntity;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
 
 import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
 
-import java.lang.reflect.Type;
 import java.util.List;
 
 /**
@@ -60,19 +56,15 @@ public class CashSalesViewModel extends AndroidViewModel {
                     public void onNetworkResponse(int type, JSONArray response, String errorMessage) {
                         switch (type){
                             case NetworkConstants.SUCCESS:
-                                DeliveryOrderEntity driverDO;
-                                try{
-                                    Type listType = new TypeToken<DeliveryOrderEntity>() {}.getType();
-                                    JSONObject obj = (JSONObject) response.get(0);
-                                    driverDO = new Gson().fromJson(String.valueOf(
-                                            obj.getJSONArray("data").get(0)), listType);
+                                DeliveryOrderParser deliveryOrderParser = new DeliveryOrderParser();
+                                List<DeliveryOrderEntity> deliveryOrdersList = deliveryOrderParser.deliveryOrdersByParsingJsonResponse(response);
+                                if(deliveryOrdersList.size() > 0){
+                                    DeliveryOrderEntity driverDO = deliveryOrdersList.get(0);
                                     leepDatabase.deliveryOrderDao().deleteDriverDo();
                                     leepDatabase.deliveryOrderDao().insertDeliveryOrder(driverDO);
-
-                                }catch (JSONException ex){
-                                    ex.printStackTrace();
                                 }
-                                    break;
+                                break;
+
                                 case NetworkConstants.UNAUTHORIZED :
                                     driverDOCallback.onResponseReceived(null,
                                             false, errorMessage, true);
@@ -82,23 +74,18 @@ public class CashSalesViewModel extends AndroidViewModel {
                 });
     }
 
-    public void fetchDriverDoDetails(final Integer doId, final UILevelNetworkCallback doDetailsCallback){
+    public void fetchDriverDoDetails(final int doId, final UILevelNetworkCallback doDetailsCallback){
         NetworkService.sharedInstance().getNetworkClient().makeGetRequest(context,
                 UrlConstants.DELIVERY_ORDERS_URL + "/" + doId, true, new NetworkAPICallback() {
                     @Override
                     public void onNetworkResponse(int type, JSONArray response, String errorMessage) {
                         switch (type){
                             case NetworkConstants.SUCCESS:
-                                try{
-                                    JSONObject obj = (JSONObject) response.get(0);
-                                    DoDetailResponseModel responseModel = new DoDetailResponseModel();
-                                    responseModel.parseJSON(obj.optJSONArray("delivery_order_items"), doId);
-                                    leepDatabase.deliveryOrderItemDao().deleteAndInsertItems(doId,
-                                            responseModel.getOrderItemEntities());
-                                }catch (JSONException ex){
-                                    ex.printStackTrace();
-                                }
+                                OrderItemParser orderItemParser = new OrderItemParser();
+                                leepDatabase.deliveryOrderItemDao().deleteAndInsertItems(doId,
+                                        orderItemParser.orderItemsByParsingJsonResponse(response, doId));
                                 break;
+
                             case NetworkConstants.UNAUTHORIZED :
                                 doDetailsCallback.onResponseReceived(null,
                                         false, errorMessage, true);
